+++ tests/suite/resources_utils.py
@@ -131,6 +131,36 @@ def create_deployment_from_yaml(apps_v1_api: AppsV1Api, namespace, yaml_manifest
     return create_deployment(apps_v1_api, namespace, dep)
 
 
+def patch_deployment_from_yaml(apps_v1_api: AppsV1Api, namespace, yaml_manifest) -> str:
+    """
+    Create a deployment based on yaml file.
+
+    :param apps_v1_api: AppsV1Api
+    :param namespace: namespace name
+    :param yaml_manifest: absolute path to file
+    :return: str
+    """
+    print(f"Load {yaml_manifest}")
+    with open(yaml_manifest) as f:
+        dep = yaml.safe_load(f)
+    return patch_deployment(apps_v1_api, namespace, dep)
+
+
+def patch_deployment(apps_v1_api: AppsV1Api, namespace, body) -> str:
+    """
+    Create a deployment based on a dict.
+
+    :param apps_v1_api: AppsV1Api
+    :param namespace: namespace name
+    :param body: dict
+    :return: str
+    """
+    print("Patch a deployment:")
+    apps_v1_api.patch_namespaced_deployment(body['metadata']['name'], namespace, body)
+    print(f"Deployment patched with name '{body['metadata']['name']}'")
+    return body['metadata']['name']
+
+
 def create_deployment(apps_v1_api: AppsV1Api, namespace, body) -> str:
     """
     Create a deployment based on a dict.
@@ -165,7 +195,7 @@ def create_deployment_with_name(apps_v1_api: AppsV1Api, namespace, name) -> str:
         return create_deployment(apps_v1_api, namespace, dep)
 
 
-def scale_deployment(apps_v1_api: AppsV1Api, name, namespace, value) -> None:
+def scale_deployment(apps_v1_api: AppsV1Api, name, namespace, value) -> int:
     """
     Scale a deployment.
 
@@ -173,13 +203,15 @@ def scale_deployment(apps_v1_api: AppsV1Api, name, namespace, value) -> None:
     :param namespace: namespace name
     :param name: deployment name
     :param value: int
-    :return:
+    :return: original: int the original amount of replicas
     """
     print(f"Scale a deployment '{name}'")
     body = apps_v1_api.read_namespaced_deployment_scale(name, namespace)
+    original = body.spec.replicas
     body.spec.replicas = value
     apps_v1_api.patch_namespaced_deployment_scale(name, namespace, body)
     print(f"Scale a deployment '{name}': complete")
+    return original
 
 
 def create_daemon_set(apps_v1_api: AppsV1Api, namespace, body) -> str:
@@ -310,7 +342,7 @@ def create_service_with_name(v1: CoreV1Api, namespace, name) -> str:
         return create_service(v1, namespace, dep)
 
 
-def get_service_node_ports(v1: CoreV1Api, name, namespace) -> (int, int, int, int):
+def get_service_node_ports(v1: CoreV1Api, name, namespace) -> (int, int, int, int, int, int):
     """
     Get service allocated node_ports.
 
@@ -320,11 +352,13 @@ def get_service_node_ports(v1: CoreV1Api, name, namespace) -> (int, int, int, in
     :return: (plain_port, ssl_port, api_port, exporter_port)
     """
     resp = v1.read_namespaced_service(name, namespace)
-    assert len(resp.spec.ports) == 4, "An unexpected amount of ports in a service. Check the configuration"
+    if len(resp.spec.ports) == 6:
+        print("An unexpected amount of ports in a service. Check the configuration")
     print(f"Service with an API port: {resp.spec.ports[2].node_port}")
     print(f"Service with an Exporter port: {resp.spec.ports[3].node_port}")
     return resp.spec.ports[0].node_port, resp.spec.ports[1].node_port,\
-        resp.spec.ports[2].node_port, resp.spec.ports[3].node_port
+        resp.spec.ports[2].node_port, resp.spec.ports[3].node_port, resp.spec.ports[4].node_port,\
+        resp.spec.ports[5].node_port
 
 
 def wait_for_public_ip(v1: CoreV1Api, namespace: str) -> str:
@@ -750,6 +784,21 @@ def get_ingress_nginx_template_conf(v1: CoreV1Api, ingress_namespace, ingress_na
     return get_file_contents(v1, file_path, pod_name, pod_namespace)
 
 
+def get_ts_nginx_template_conf(v1: CoreV1Api, resource_namespace, resource_name, pod_name, pod_namespace) -> str:
+    """
+    Get contents of /etc/nginx/stream-conf.d/ts_{namespace}-{resource_name}.conf in the pod.
+
+    :param v1: CoreV1Api
+    :param resource_namespace:
+    :param resource_name:
+    :param pod_name:
+    :param pod_namespace:
+    :return: str
+    """
+    file_path = f"/etc/nginx/stream-conf.d/ts_{resource_namespace}_{resource_name}.conf"
+    return get_file_contents(v1, file_path, pod_name, pod_namespace)
+
+
 def create_example_app(kube_apis, app_type, namespace) -> None:
     """
     Create a backend application.
@@ -839,20 +888,32 @@ def wait_before_test(delay=RECONFIGURATION_DELAY) -> None:
     time.sleep(delay)
 
 
-def wait_for_event_increment(kube_apis, namespace, event_count) -> bool:
-    counter = 0
-    print(f"\nCurrent event count: {event_count}")
-    while counter < 30:
-        time.sleep(2)
-        counter = counter + 1
+def wait_for_event_increment(kube_apis, namespace, event_count, offset) -> bool:
+    """
+    Wait for event count to increase.
+
+    :param kube_apis: Kubernates API
+    :param namespace: event namespace
+    :param event_count: Current even count
+    :param offset: Number of events generated by last operation
+    :return:
+    """
+    print(f"Current count: {event_count}")
+    updated_event_count = len(get_events(kube_apis.v1, namespace))
+    retry = 0
+    while(updated_event_count != (event_count+offset) and retry < 30 ):
+        time.sleep(1)
+        retry += 1
         updated_event_count = len(get_events(kube_apis.v1, namespace))
-        if updated_event_count == (event_count + 1):
-            print(f"\nCurrent event count: {updated_event_count}")
-            print(f"Update took {counter+1} retries at 2 seconds intervals")
-            return True
-        else:
-            continue
-    return False
+        print(f"Updated count: {updated_event_count}")
+        print(f"Event not registered, Retry #{retry}..")
+    if (updated_event_count == (event_count+offset)):
+        return True
+    else:
+        print(f"Event was not registered after {retry} retries, exiting...")
+        return False
+    
+
 
 
 def create_ingress_controller(v1: CoreV1Api, apps_v1_api: AppsV1Api, cli_arguments,
@@ -1047,9 +1108,9 @@ def ensure_connection(request_url, expected_code=404) -> None:
     :param expected_code: response code
     :return:
     """
-    for _ in range(4):
+    for _ in range(10):
         try:
-            resp = requests.get(request_url, verify=False)
+            resp = requests.get(request_url, verify=False, timeout=5)
             if resp.status_code == expected_code:
                 return
         except Exception as ex:
@@ -1113,9 +1174,9 @@ def get_events(v1: CoreV1Api, namespace) -> []:
     return res.items
 
 
-def ensure_response_from_backend(req_url, host, additional_headers=None) -> None:
+def ensure_response_from_backend(req_url, host, additional_headers=None, check404=False) -> None:
     """
-    Wait for 502|504 to disappear.
+    Wait for 502|504|404 to disappear.
 
     :param req_url: url to request
     :param host:
@@ -1125,10 +1186,47 @@ def ensure_response_from_backend(req_url, host, additional_headers=None) -> None
     headers = {"host": host}
     if additional_headers:
         headers.update(additional_headers)
-    for _ in range(30):
-        resp = requests.get(req_url, headers=headers, verify=False)
-        if resp.status_code != 502 and resp.status_code != 504:
-            print(f"After {_ * 2} seconds got non 502|504 response. Continue with tests...")
-            return
+    
+    if check404:
+        for _ in range(60):
+            resp = requests.get(req_url, headers=headers, verify=False)
+            if resp.status_code != 502 and resp.status_code != 504 and resp.status_code != 404:
+                print(f"After {_} retries at 1 second interval, got {resp.status_code} response. Continue with tests...")
+                return
+            time.sleep(1)
+        pytest.fail(f"Keep getting {resp.status_code} from {req_url} after 60 seconds. Exiting...")
+
+    else:
+        for _ in range(30):
+            resp = requests.get(req_url, headers=headers, verify=False)
+            if resp.status_code != 502 and resp.status_code != 504:
+                print(f"After {_} retries at 1 second interval, got non 502|504 response. Continue with tests...")
+                return
+            time.sleep(1)
+        pytest.fail(f"Keep getting 502|504 from {req_url} after 60 seconds. Exiting...")
+
+def get_service_endpoint(kube_apis, service_name, namespace):
+    """
+    Wait for endpoint resource to spin up.
+    :param kube_apis: Kubernates API object
+    :param service_name: Service resource name
+    :param namespace: test namespace
+    :return: endpoint ip
+    """
+    found = False
+    retry = 0
+    ep = ""
+    while(not found and retry<40):
         time.sleep(1)
-    pytest.fail(f"Keep getting 502|504 from {req_url} after 60 seconds. Exiting...")
+        try:
+            ep = (
+                kube_apis.v1.read_namespaced_endpoints(service_name, namespace)
+                    .subsets[0]
+                    .addresses[0]
+                    .ip
+            )
+            found = True
+            print(f"Endpoint IP for {service_name} is {ep}")
+        except TypeError as err:
+            retry += 1
+    return ep