+++ docs-web/releases.md
@@ -1,5 +1,35 @@
 # Releases
 
+## NGINX Ingress Controller 1.11.3
+
+25 May 2021
+
+CHANGES:
+* Update NGINX version to 1.21.0.
+
+HELM CHART:
+* The version of the Helm chart is now 0.9.3.
+
+UPGRADE:
+* For NGINX, use the 1.11.3 image from our DockerHub: `nginx/nginx-ingress:1.11.3`, `nginx/nginx-ingress:1.11.3-alpine` or `nginx/nginx-ingress:1.11.3-ubi`
+* For NGINX Plus, please build your own image using the 1.11.3 source code.
+* For Helm, use version 0.9.3 of the chart.
+
+## NGINX Ingress Controller 1.11.2
+
+19 May 2021
+
+CHANGES:
+* Update NGINX Plus version to R23 P1.
+
+HELM CHART:
+* The version of the Helm chart is now 0.9.2.
+
+UPGRADE:
+* For NGINX, use the 1.11.2 image from our DockerHub: `nginx/nginx-ingress:1.11.2`, `nginx/nginx-ingress:1.11.2-alpine` or `nginx/nginx-ingress:1.11.2-ubi`
+* For NGINX Plus, please build your own image using the 1.11.2 source code.
+* For Helm, use version 0.9.2 of the chart.
+
 ## NGINX Ingress Controller 1.11.1
 
 7 April 2021
@@ -79,7 +109,7 @@ SUPPORTED PLATFORMS:
 
 We will provide technical support for the NGINX Ingress Controller on any Kubernetes platform that is currently supported by its provider and which passes the Kubernetes conformance tests.  This release was fully tested on the following Kubernetes versions: 1.16-1.20.
 
-### NGINX Ingress Controller 1.10.1
+## NGINX Ingress Controller 1.10.1
 
 16 March 2021